+++ resources/views/pages/partials/blocks/image.blade.php
@@ -14,18 +14,51 @@
   $resolvedAltText = $altText !== '' ? $altText : $fallbackAltText;
   $href = trim((string) $block->boundPublicValue('url', $block->url ?? ''));
   $linkAttributes = '';
+  $viewerEnabled = (bool) $block->setting('viewer_enabled', false);
+  $viewerGroup = trim((string) $block->setting('viewer_group', 'page-images')) ?: 'page-images';
+  $viewerRegistry = app(\WebBlocks\Cms\Support\Blocks\PublicOverlayRegistry::class);
+  $viewerId = $viewerRegistry->imageViewerId($viewerGroup);
+  $fullImageSource = $image?->url() ?: $imageSource;
 
   if ($href !== '' && preg_match('/^(https?:\/\/|\/|#|mailto:|tel:)/i', $href)) {
     $linkAttributes = ' href="'.e($href).'"';
   } else {
     $href = '';
   }
+
+  $opensViewer = $viewerEnabled && $href === '' && $fullImageSource;
+
+  if ($opensViewer) {
+    $viewerRegistry->registerImageViewerItem($viewerGroup, (int) $block->id, [
+      'thumbnail_url' => $imageSource,
+      'full_url' => $fullImageSource,
+      'alt' => $resolvedAltText,
+      'caption' => $caption,
+      'meta' => '',
+      'width' => $image?->width,
+      'height' => $image?->height,
+      'srcset' => $srcset,
+    ], $block->renderLocaleCode());
+  }
 @endphp
 
 @if ($imageSource)
   <figure class="wb-stack wb-gap-2" data-wb-public-block-type="{{ $block->publicBlockTypeAttribute() }}">
-    @if ($href !== '')
-      <a{!! $linkAttributes !!}>
+    @if ($href !== '' || $opensViewer)
+      <a
+        @if ($opensViewer)
+          href="{{ $fullImageSource }}"
+          class="wb-gallery-trigger"
+          data-wb-gallery-target="#{{ $viewerId }}"
+          data-wb-gallery-full="{{ $fullImageSource }}"
+          data-wb-gallery-alt="{{ $resolvedAltText }}"
+          @if ($caption !== '') data-wb-gallery-caption="{{ $caption }}" @endif
+          @if ($image?->width) data-wb-gallery-width="{{ $image->width }}" @endif
+          @if ($image?->height) data-wb-gallery-height="{{ $image->height }}" @endif
+        @else
+          {!! $linkAttributes !!}
+        @endif
+      >
     @endif
     <img
       src="{{ $imageSource }}"
@@ -36,7 +69,7 @@
       @if ($image?->width) width="{{ $image->width }}" @endif
       @if ($image?->height) height="{{ $image->height }}" @endif
     >
-    @if ($href !== '')
+    @if ($href !== '' || $opensViewer)
       </a>
     @endif